# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.EmployeeRequest:
    properties:
      email:
        type: string
      employeeCode:
        type: string
      employmentStatus:
        type: string
      fullName:
        type: string
      positionId:
        type: string
      unitId:
        type: string
    type: object
...
  dto.PositionRequest:
    properties:
      companyId:
        type: string
      description:
        type: string
      level:
        type: string
      title:
        type: string
    type: object
...
  dto.UnitRequest:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
...
  model.Employee:
    properties:
      createdAt:
        type: string
      createdById:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      employeeCode:
        type: string
      employmentStatus:
        $ref: '#/definitions/model.EmploymentStatus'
      endDate:
        type: string
      fullName:
        type: string
      id:
        type: string
      phone:
        type: string
      position:
        $ref: '#/definitions/model.Position'
      positionId:
        type: string
      startDate:
        type: string
      unit:
        $ref: '#/definitions/model.Unit'
      unitId:
        type: string
      updatedAt:
        type: string
      updatedById:
        type: string
    type: object
  model.EmploymentStatus:
    enum:
    - FULLTIME
    - CONTRACT
    - INTERN
    - PARTTIME
    type: string
    x-enum-varnames:
    - EmploymentFullTime
    - EmploymentContract
    - EmploymentIntern
    - EmploymentPartTime
...
  model.Position:
    properties:
      companyId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      level:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  model.Unit:
    properties:
      code:
        type: string
      createdAt:
        type: string
      createdById:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      updatedById:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Human Resource Information System API
  title: HRIS API
  version: "1.0"
paths:
...
  /master/employees:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Employee'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all employees
      tags:
      - master-data
    post:
      consumes:
      - application/json
      description: Create a new employee in the system (Admin, HR, Manager only)
      parameters:
      - description: Employee data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new employee
      tags:
      - master-data
  /master/employees/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an employee from the system (Admin, HR, Manager only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an employee
      tags:
      - master-data
    get:
      consumes:
      - application/json
      description: Retrieve a specific employee by their ID (Admin, HR, Manager only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get employee by ID
      tags:
      - master-data
    put:
      consumes:
      - application/json
      description: Update an existing employee's information (Admin, HR, Manager only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated employee data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - master-data
  /master/positions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all positions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Position'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all positions
      tags:
      - master-data
    post:
      consumes:
      - application/json
      description: Create a new position in the system (Admin, HR only)
      parameters:
      - description: Position data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PositionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Position'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new position
      tags:
      - master-data
  /master/units:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Unit'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all units
      tags:
      - master-data
    post:
      consumes:
      - application/json
      description: Create a new unit in the system (Admin, HR only)
      parameters:
      - description: Unit data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Unit'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new unit
      tags:
      - master-data
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"